+++ Changelog.md
@@ -2,6 +2,7 @@
 
 - Add prop to change canvas background color
 - Move some props to options
+- fix Picture shape not being saved and not triggering onDataChanged
 
 # V.0.3.0 :